# improved performance of CalculateComfortLevel