You already made `GetAllPiecesInRadius` a lot faster. It is currently only used for `CalculateComfortLevel`. With your `GetAllPiecesInRadius` already active here are a few figures inside my world:

Main Base: ~20ms => ~5ms
Treehouse: ~35ms => ~6ms
Tower: ~70ms => ~10ms

You should verify that the comfort level is correctly calculated.