@ontrigger 2021-06-21
Can confirm the improvement, the spikes went from 18ms to just 2. 
The comfort level seems to match with the original, lgtm